Conflicts:
	cars/build.gradle
	cars/src/main/java/dk/aau/cs/giraf/cars/DatabaseHelper.java

+++ cars/src/main/java/dk/aau/cs/giraf/cars/SettingsActivity.java
@@ -33,7 +33,7 @@ public class SettingsActivity extends Activity {
 
         Intent intent = getIntent();
 
-        int guardianId = 0;
+        long guardianId = 0;
         DatabaseHelper database = new DatabaseHelper(this);
 
         if(intent.hasExtra(DatabaseHelper.CHILD_ID))
